+++ nautilus_core/common/src/clock.rs
@@ -27,11 +27,11 @@ use pyo3::AsPyPointer;
 
 #[pyclass]
 pub struct TestClock {
-    time_ns: Timestamp,
-    next_time_ns: Timestamp,
-    timers: HashMap<String, TestTimer>,
-    handlers: HashMap<String, PyObject>,
-    default_handler: PyObject,
+    pub time_ns: Timestamp,
+    pub next_time_ns: Timestamp,
+    pub timers: HashMap<String, TestTimer>,
+    pub handlers: HashMap<String, PyObject>,
+    pub default_handler: PyObject,
 }
 
 impl TestClock {
@@ -66,7 +66,7 @@ impl TestClock {
     }
 
     #[inline]
-    fn advance_time(&mut self, to_time_ns: Timestamp) -> Vec<TimeEventHandler> {
+    pub fn advance_time(&mut self, to_time_ns: Timestamp) -> Vec<TimeEventHandler> {
         // Time should increase monotonically
         assert!(
             to_time_ns >= self.time_ns,
@@ -180,14 +180,14 @@ pub struct TimeEventHandler {
 
 impl TimeEventHandler {
     #[inline]
-    fn handle_py(self) {
+    pub fn handle_py(self) {
         Python::with_gil(|py| {
             let _ = self.handler.call0(py);
         });
     }
 
     #[inline]
-    fn handle(self) {
+    pub fn handle(self) {
         Python::with_gil(|py| {
             let _ = self.handler.call1(py, (self.event,));
         });
@@ -250,83 +250,4 @@ pub extern "C" fn advance_time(clock: &mut CTestClock, to_time_ns: u64) -> PyObj
     Python::with_gil(|py| {
         PyList::new(py, events.into_iter().map(|v| Py::new(py, v).unwrap())).into()
     })
-}
-
-#[cfg(test)]
-mod tests {
-    use std::str::FromStr;
-
-    use pyo3::{prelude::*, types::*};
-
-    use crate::clock::set_time_alert_ns;
-
-    use super::new_test_clock;
-
-    #[test]
-    fn test_test_clock() {
-        pyo3::prepare_freethreaded_python();
-        let mut test_clock = Python::with_gil(|py| {
-            let dummy = PyDict::new(py).into();
-            new_test_clock(0, dummy)
-        });
-
-        assert_eq!(test_clock.time_ns, 0);
-        let timer_name = "tringtring";
-
-        let (name, callback) = Python::with_gil(|py| {
-            let name = PyString::new(py, timer_name).into();
-            let dummy = Some(PyDict::new(py).into());
-            (name, dummy)
-        });
-
-        unsafe {
-            set_time_alert_ns(&mut test_clock, name, 2_000, callback);
-        }
-
-        assert_eq!(test_clock.timers.len(), 1);
-        assert_eq!(
-            test_clock.timers.keys().next().unwrap().as_str(),
-            timer_name
-        );
-
-        let events = test_clock.advance_time(3_000);
-
-        assert_eq!(test_clock.timers.values().next().unwrap().is_expired, true);
-        assert_eq!(events.len(), 1);
-        assert_eq!(
-            events.iter().next().unwrap().event.name.to_string(),
-            String::from_str(timer_name).unwrap()
-        );
-    }
-
-    #[test]
-    fn test_event_callback() {
-        pyo3::prepare_freethreaded_python();
-        let mut test_clock = Python::with_gil(|py| {
-            let dummy = PyDict::new(py).into();
-            new_test_clock(0, dummy)
-        });
-
-        let (name, callback, pymod): (PyObject, PyObject, PyObject) = Python::with_gil(|py| {
-            let code = include_str!("./test_data/callback.py");
-            let pymod = PyModule::from_code(py, &code, "humpty", "dumpty").unwrap();
-            let name = PyString::new(py, "brrrringbrrring");
-            let callback = pymod.getattr("increment").unwrap();
-            (name.into(), callback.into(), pymod.into())
-        });
-
-        unsafe {
-            set_time_alert_ns(&mut test_clock, name, 2_000, Some(callback));
-        }
-
-        let events = test_clock.advance_time(3_000);
-        events
-            .into_iter()
-            .for_each(|time_event_handler| time_event_handler.handle());
-
-        let count: u64 =
-            Python::with_gil(|py| pymod.getattr(py, "count").unwrap().extract(py).unwrap());
-
-        assert_eq!(count, 1);
-    }
 }
